Fix dungeoneering circular reference
