+++ src/commands/Minion/dung.ts
@@ -1,173 +1,29 @@
 import { reduceNumByPercent, Time } from 'e';
-import { CommandStore, KlasaMessage, KlasaUser } from 'klasa';
+import { CommandStore, KlasaMessage } from 'klasa';
 
 import { Emoji } from '../../lib/constants';
 import { minionNotBusy, requiresMinion } from '../../lib/minions/decorators';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
+import {
+	determineDgLevelForFloor,
+	dungBuyables,
+	DungeoneeringOptions,
+	isValidFloor,
+	requiredSkills
+} from '../../lib/skilling/skills/dung/dungData';
+import {
+	gorajanShardChance,
+	hasRequiredLevels,
+	maxFloorUserCanDo,
+	numberOfGorajanOutfitsEquipped
+} from '../../lib/skilling/skills/dung/dungDbFunctions';
 import { SkillsEnum } from '../../lib/skilling/types';
 import { BotCommand } from '../../lib/structures/BotCommand';
 import { MakePartyOptions } from '../../lib/types';
-import { ActivityTaskOptions } from '../../lib/types/minions';
-import { formatDuration, formatSkillRequirements, skillsMeetRequirements, stringMatches } from '../../lib/util';
+import { formatDuration, formatSkillRequirements, stringMatches } from '../../lib/util';
 import addSubTaskToActivityTask from '../../lib/util/addSubTaskToActivityTask';
 import { formatOrdinal } from '../../lib/util/formatOrdinal';
-import getOSItem from '../../lib/util/getOSItem';
 import itemID from '../../lib/util/itemID';
-import { gorajanShardChance, numberOfGorajanOutfitsEquipped } from '../../tasks/minions/dungeoneeringActivity';
-
-export type Floor = 1 | 2 | 3 | 4 | 5 | 6 | 7;
-
-export function isValidFloor(floor: number | string): floor is Floor {
-	return [1, 2, 3, 4, 5, 6, 7].includes(floor as number);
-}
-
-export interface DungeoneeringOptions extends ActivityTaskOptions {
-	leader: string;
-	users: string[];
-	quantity: number;
-	floor: number;
-}
-
-const dungBuyables = [
-	{
-		item: getOSItem('Chaotic rapier'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Chaotic longsword'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Chaotic maul'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Chaotic staff'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Chaotic crossbow'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Offhand Chaotic rapier'),
-		cost: 100_000
-	},
-	{
-		item: getOSItem('Offhand Chaotic longsword'),
-		cost: 100_000
-	},
-	{
-		item: getOSItem('Offhand Chaotic crossbow'),
-		cost: 100_000
-	},
-	{
-		item: getOSItem('Farseer kiteshield'),
-		cost: 200_000
-	},
-	{
-		item: getOSItem('Scroll of life'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Herbicide'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of efficiency'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of farming'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of cleansing'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of dexterity'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of teleportation'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of mystery'),
-		cost: 500_000
-	},
-	{
-		item: getOSItem('Amulet of zealots'),
-		cost: 400_000
-	},
-	{
-		item: getOSItem('Scroll of proficiency'),
-		cost: 900_000
-	},
-	{
-		item: getOSItem('Frosty'),
-		cost: 2_000_000
-	},
-	{
-		item: getOSItem('Chaotic remnant'),
-		cost: 500_000
-	},
-	{
-		item: getOSItem('Scroll of longevity'),
-		cost: 800_000
-	},
-	{
-		item: getOSItem('Scroll of the hunt'),
-		cost: 800_000
-	},
-	{
-		item: getOSItem('Daemonheim agility pass'),
-		cost: 1_000_000
-	},
-	{
-		item: getOSItem('Dungeoneering dye'),
-		cost: 4_000_000
-	}
-];
-
-function determineDgLevelForFloor(floor: number) {
-	return Math.floor(floor * 20 - 20);
-}
-
-function requiredLevel(floor: number) {
-	return floor * 14;
-}
-
-function requiredSkills(floor: number) {
-	const lvl = requiredLevel(floor);
-	const nonCmbLvl = Math.floor(lvl / 1.5);
-	return {
-		attack: lvl,
-		strength: lvl,
-		defence: lvl,
-		hitpoints: lvl,
-		magic: lvl,
-		ranged: lvl,
-		herblore: nonCmbLvl,
-		runecraft: nonCmbLvl,
-		prayer: nonCmbLvl,
-		fletching: nonCmbLvl,
-		fishing: nonCmbLvl,
-		cooking: nonCmbLvl,
-		construction: nonCmbLvl,
-		crafting: nonCmbLvl,
-		dungeoneering: determineDgLevelForFloor(floor)
-	};
-}
-
-function hasRequiredLevels(user: KlasaUser, floor: number) {
-	return skillsMeetRequirements(user.rawSkills, requiredSkills(floor));
-}
-
-export function maxFloorUserCanDo(user: KlasaUser) {
-	return [7, 6, 5, 4, 3, 2, 1].find(floor => hasRequiredLevels(user, floor)) || 1;
-}
 
 // Max people in a party:
 const maxTeamSize = 20;
